Activation Functions from NNlib.jl

These non-linearities used between layers of your model are exported by the NNlib package.

Note that, unless otherwise stated, activation functions operate on scalars. To apply them to an array you can call Пғ.(xs), relu.(xs) and so on. Alternatively, they can be passed to a layer like Dense(784 => 1024, relu) which will handle this broadcasting.

Functions like softmax are sometimes described as activation functions, but not by Flux. They must see all the outputs, and hence cannot be broadcasted. See the next page for details.

Alphabetical Listing

NNlib.celu вҖ” Function

celu(x, Оұ=1) = x вүҘ 0 ? x : Оұ * (exp(x/Оұ) - 1)

Activation function from "Continuously Differentiable Exponential Linear Units".

julia> lineplot(celu, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв ’в үв”Ӯ celu(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв үв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЈҖв Өв –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎӨвЎв ¶в ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЈҖв Өв ”в ’в Ӣв Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -1 в”Ӯв Өв Өв Өв Өв ”в ’в ’в ’в Ҡв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

julia> celu(-10f0)
-0.9999546f0

NNlib.elu вҖ” Function

elu(x, Оұ=1) = x > 0 ? x : Оұ * (exp(x) - 1)

Exponential Linear Unit activation function. See "Fast and Accurate Deep Network Learning by Exponential Linear Units". You can also specify the coefficient explicitly, e.g. elu(x, 1).

julia> lineplot(elu, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ       
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв ’в үв”Ӯ elu(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв үв Җв Җв Җв Җв”Ӯ       
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЈҖв Өв –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
           в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎӨвЎв ¶в ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ       
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЈҖв Өв ”в ’в Ӣв Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
        -1 в”Ӯв Өв Өв Өв Өв ”в ’в ’в ’в Ҡв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ       
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ       
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ       

julia> elu(-10f0)
-0.9999546f0

julia> elu(-10f0, 2)
-1.9999092f0

NNlib.gelu вҖ” Function

gelu(x) = 0.5x * (1 + tanh(вҲҡ(2/ПҖ) * (x + 0.044715x^3)))

Activation function from "Gaussian Error Linear Units".

julia> lineplot(gelu, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв”Ӯ gelu(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ҡв Ғв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв ’в үв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв ҖвЈҖвЎ в Өв ’в үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”ӮвЈӨвЈӨвЈӨвЈӨвЈӨвЈӨвЈӨвЈӨвЎӨв Өв Өв Өв Өв Өв Өв ӨвЈӨвЈӨвЈӨвЎӨвЎв ¶в ¶в ӯв Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв үв үв үв үв үв үв үв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

julia> lineplot(gelu, -5, 0, height=7);

julia> lineplot!(ans, swish)
             в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
           0 в”Ӯв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв ’в ’в ӨвЈ„вЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв”Ӯ gelu(x) 
             в”Ӯв ‘в ’в ўв ӨвЈ„вЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв “вў„вЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв”Ӯ swish(x)
             в”Ӯв Җв Җв Җв Җв Җв Ҳв үв ’в ӨвЈҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ‘вўҶвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈёв Ғв”Ӯ         
   f(x)      в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв ’вў„вЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ‘вў„в Җв Җв Җв Җв Җв Җв Җв Җвў вЎҮв Җв”Ӯ         
             в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв “вў„вЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв “вЈ„в Җв Җв Җв Җв Җвў вЎһв Җв Җв”Ӯ         
             в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв ҰвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв “вў„вЈҖвЈҖвЎӨвўЈв ғв Җв Җв”Ӯ         
        -0.2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв “вЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвў в Үв Җв Җв Җв”Ӯ         
             в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
             в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ0в Җ         
             в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

NNlib.hardsigmoid вҖ” Function

hardПғ(x) = max(0, min(1, (x + 3) / 6))

Piecewise linear approximation of sigmoid.

julia> lineplot(hardsigmoid, -5, 5, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
        1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в –в Ӣв үв үв үв үв үв үв үв үв”Ӯ hardПғ(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвЈҖвЎӨв ’в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв ҖвЎ в ”в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨвЎ—в үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ҡв Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв –в Ӣв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв Өв Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
          в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ         
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

julia> lineplot(sigmoid, -5, 5, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ     
        1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖвЎ в Өв –в ’в ’в Ӣв үв үв үв үв үв үв”Ӯ Пғ(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвўҖвЎ в –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвЈҖв ”в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ вЎҸв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ”в Ӣв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв Ҡв Ғв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв Өв Өв Өв ’в Ҡв үв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ     
          в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ     
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ     

NNlib.hardswish вҖ” Function

hardswish(x) = x * hardПғ(x)

Hard-Swish activation function. See "Searching for MobileNetV3".

julia> lineplot(hardswish, -2, 5, height = 7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ             
         5 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв ”в ’в үв”Ӯ hardswish(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎ в ”в ’в үв Ғв Җв Җв Җв Җв”Ӯ             
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв –в үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв ҖвўҖвЈҖв Өв –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
           в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҮвЈӨвЈӨвЈ–вЈҡвЈүвЈҒвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв”Ӯ             
        -1 в”Ӯв үв ’в ’в ’в ’в үв үв үв үв Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ             
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ             
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ             

julia> lineplot(hardswish, -4, 0, height = 7);

julia> lineplot!(ans, swish)
             в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ             
           0 в”Ӯв үв үв үв үв үв үв үв үв үв үвўЈвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎңв”Ӯ hardswish(x)
             в”Ӯв ’в ’в ўв Өвў„вЈҖвЎҖв Җв Җв Җв Җв ұвЎ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖв Һв Җв”Ӯ swish(x)    
             в”Ӯв Җв Җв Җв Җв Җв Җв Ҳв үв ‘в ’в Ұвў„вЈҳвў„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎҙв ғв Җв Җв”Ӯ             
   f(x)      в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв ‘вЎ–в Ұвў„вЈҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвў”в Ҹв Ғв Җв Җв Җв”Ӯ             
             в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв ЈвЈ„в Җв үв ‘в ’в Ұв Өвў„вЈҖвЈҖвЈҖвЈҖвЎ в Өв –вЈҠв •в Ғв Җв Җв Җв Җв Җв”Ӯ             
             в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв “в ӨвЎҖв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв –в Ғв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
        -0.4 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв үв ’в ўв Өв Өв ”в ’в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ             
             в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ             
             в Җ-4в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ0в Җ             
             в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ             

julia> hardswish.(-5:5)'
1Г—11 adjoint(::Vector{Float64}) with eltype Float64:
 -0.0  -0.0  -0.0  -0.333333  -0.333333  0.0  0.666667  1.66667  3.0  4.0  5.0

NNlib.hardtanh вҖ” Function

hardtanh(x) = max(-1, min(1, x))

Segment-wise linear approximation of tanh, much cheaper to compute. See "Large Scale Machine Learning".

See also tanh_fast.

julia> lineplot(hardtanh, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ            
         1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ӣв үв үв үв үв үв үв үв үв үв үв”Ӯ hardtanh(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв ҖвЈҖвЎӨв Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвўҖвЎӨв Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎӨвЎ·в Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в –в Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в –в Ӣв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
        -1 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв ”в Ӣв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ            
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ            
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җx

julia> lineplot(tanh, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎ в Өв Өв ’в ’в ’в Ҡв үв үв үв”Ӯ tanh(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв ҖвўҖвЎ в ”в Ҡв үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвўҖвЎӨв ’в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎӨвЎ·в Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв –в Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎ в ”в Ҡв Ғв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -1 в”ӮвЈҖвЈҖвЈҖвЎ в Өв Өв Өв –в ’в Ҡв үв Ғв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

NNlib.leakyrelu вҖ” Function

leakyrelu(x, a=0.01) = max(a*x, x)

Leaky Rectified Linear Unit activation function. You can also specify the coefficient explicitly, e.g. leakyrelu(x, 0.01).

julia> lineplot(x -> leakyrelu(x, 0.5), -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ       
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв ’в үв”Ӯ #42(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв үв Җв Җв Җв Җв”Ӯ       
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЈҖв Өв –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
           в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвЈӨвЈӨвЎӨвЎв ¶в ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ       
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЈҖв Өв Өв ’в ’в Ӣв үв Ғв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
        -1 в”ӮвЈҖвЈҖв Өв Өв ’в ’в Ҡв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ       
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ       
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ       
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ       

julia> leakyrelu(-10f0, 0.2)
-2.0f0

julia> leakyrelu(-10f0, 0.02)
-0.5f0

NNlib.lisht вҖ” Function

lisht(x) = x * tanh(x)

Activation function from "LiSHT: Non-Parametric Linearly Scaled Hyperbolic Tangent ..."

julia> lineplot(lisht, -2, 2, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
        2 в”Ӯв ўвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в”Ӯ lisht(x)
          в”Ӯв Җв Ҳв ‘вўҰвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв Ҡв Ғв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Ҳв ЈвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ғв Җв Җв Җв Җв”Ӯ         
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв ‘вўҶвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в Ҡв Ғв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв ўвЎ„в Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖв ”в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв “вў„вЎҖв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвўҖвЎ в –в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        0 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв “в ҰвЈ„вЈҖвЈҖвЈҮвЈҖвЈҖв Өв ’в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
          в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ         
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

julia> lineplot!(ans, logcosh)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ           
        2 в”Ӯв ўвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в”Ӯ lisht(x)  
          в”Ӯв Җв Ҳв ‘вўҰвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв Ҡв Ғв Җв”Ӯ logcosh(x)
          в”Ӯв ўвЈ„в Җв Җв Ҳв ЈвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ғв Җв ҖвЈҖв ”в”Ӯ           
   f(x)   в”Ӯв Җв Ҳв ‘в ўвЈҖв Җв Җв ‘вўҶвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в Ҡв Ғв ҖвЈҖв ”в Ҡв Ғв Җв”Ӯ           
          в”Ӯв Җв Җв Җв Җв Җв үв ўвў„вЎҖв үв ўвЎ„в Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖв ”в Ӣв ҖвЎ в ”в Ӣв Ғв Җв Җв Җв Җв”Ӯ           
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв үв “в ҰвЈҢвЎ“вў„вЎҖв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвўҖвЎ в –вЈҒв Өв ’в үв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ           
        0 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв үв “в Әв ·вЈҰвЈ„вЈҖвЈҖвЈҮвЈҖвЈҖвЈӨв ¶в •в ’в үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ           
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ           
          в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ           
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ           

NNlib.logcosh вҖ” Function

logcosh(x)

Return log(cosh(x)) which is computed in a numerically stable way.

julia> lineplot(logcosh, -5, 5, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ           
        5 в”ӮвЎҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ logcosh(x)
          в”Ӯв үв ўвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ӣв”Ӯ           
          в”Ӯв Җв Җв Җв ‘в ўвЈ„в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ҡв Ғв Җв Җв”Ӯ           
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв ‘в ҰвЈҖв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв Ҡв Ғв Җв Җв Җв Җв Җв”Ӯ           
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв ‘в ҰвЎҖв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ           
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв “в ҰвЎҖв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвўҖвЎӨв ’в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ           
        0 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Ҳв ‘в ўвў„вЈҖвЈҖвЈҮвЈҖвЎ в ”в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ           
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ           
          в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ           
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ           

NNlib.logsigmoid вҖ” Function

logПғ(x)

Return log(Пғ(x)) which is computed in a numerically stable way.

julia> lineplot(logsigmoid, -5, 5, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         0 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎв Өв ”в ’в ’в ’в үв үв үв үв үв үв үв үв үв үв үв үв үв үв”Ӯ logПғ(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ҡв үв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв ’в үв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Җв Җв Җв Җв Җв ҖвўҖвЎӨв –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв ҖвЈҖв ”в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”ӮвЎӨв –в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -6 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

NNlib.mish вҖ” Function

mish(x) = x * tanh(softplus(x))

Activation function from "Mish: A Self Regularized Non-Monotonic Neural Activation Function".

julia> lineplot(mish, -5, 5, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         5 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в –в Ӣв”Ӯ mish(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв ’в Ғв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в ”в Ӣв Ғв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв ҖвўҖвЎ в –в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвўҖвЎӨв –в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈвЈ”вЈҠвЈҒвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв”Ӯ        
        -1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв үв үв үв үв үв үв үв үв үв үв үв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

NNlib.relu вҖ” Function

relu(x) = max(0, x)

Rectified Linear Unit activation function.

julia> lineplot(relu, -2, 2, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
        2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в Ӣв”Ӯ relu(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв Ҡв Ғв Җв Җв”Ӯ        
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв Ҡв Ғв Җв Җв Җв Җв Җв”Ӯ        
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖвЎӨв –в Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв ҖвЎ в –в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЎ в –в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҮв ”в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
          в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ        
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

NNlib.relu6 вҖ” Function

relu6(x) = min(max(0, x), 6)

Rectified Linear Unit activation function capped at 6. See "Convolutional Deep Belief Networks" from CIFAR-10.

julia> lineplot(relu6, -10, 10, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
        6 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Һв үв үв үв үв үв үв үв үв”Ӯ relu6(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ”в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЎӨв ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвЎ в Һв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвўҖв –в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв ҖвЎ”в ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЎв Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
          в Җ-10в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ10в Җ         
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

NNlib.rrelu вҖ” Function

rrelu(x, lo=1/8, hi=1/3) = max(a*x, x)
# where `a` is randomly sampled from uniform distribution `U(lo, hi)`

Randomized Leaky Rectified Linear Unit activation function. See "Empirical Evaluation of Rectified Activations" You can also specify the bound explicitly, e.g. rrelu(x, 0.0, 1.0).

julia> lineplot(rrelu, -20, 10, height=7)
            в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
         10 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ӣв”Ӯ rrelu(x)
            в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв Җв Җв Җв Җв ҖвўҖвЎ в –в Ӣв Ғв Җв Җв Җв”Ӯ         
            в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв ҖвўҖвЎ в ”в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
   f(x)     в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎӨв ӨвЈӨвЈӨвўӨвЈӨвЈӨв Өв Өв Өвўјв ®в Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ         
            в”ӮвЈ°вўҖвЈҶвЎ„вЈ„вЎ„вЎ вЎ°в Ұв ·вЎңвўўв ·в ів ўв Ҡв үв үв Җв Җв Ғв Җв Җв Җв Җв Җвўёв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
            в”Ӯв ғв үв ҷв ҳв ғв Ҳв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        -10 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җвўёв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
            в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
            в Җ-20в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ10в Җ         
            в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

julia> extrema(rrelu.(fill(-10f0, 1000)))
(-3.3316886f0, -1.2548422f0)

NNlib.selu вҖ” Function

selu(x) = О» * (x вүҘ 0 ? x : Оұ * (exp(x) - 1))

О» вүҲ 1.05070...
Оұ вүҲ 1.67326...

Scaled exponential linear units. See "Self-Normalizing Neural Networks".

julia> lineplot(selu, -3, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         3 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ selu(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв Өв ’в”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв ҖвўҖвЈҖв Өв –в Ҡв үв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвЈҖвЎ в Өв ’в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үвЈүв ӯв ӣвЎҸв үв үв үв үв үв үв үв үв үв үв үв үв үв үв үв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЈҖвЎӨв Өв ’в Ҡв үв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -2 в”Ӯв Өв Өв –в ’в ’в ’в ’в ’в ’в ’в үв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-3в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ        

julia> selu(-10f0)
-1.7580194f0

NNlib.sigmoid вҖ” Function

Пғ(x) = 1 / (1 + exp(-x))

Classic sigmoid activation function. Unicode Пғ can be entered as \sigma then tab, in many editors. The ascii name sigmoid is also exported.

See also sigmoid_fast.

julia> using UnicodePlots

julia> lineplot(sigmoid, -5, 5, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ     
        1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЈҖвЎ в Өв –в ’в ’в Ӣв үв үв үв үв үв үв”Ӯ Пғ(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвўҖвЎ в –в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвЈҖв ”в Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ вЎҸв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ”в Ӣв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв Ҡв Ғв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖв Өв Өв Өв ’в Ҡв үв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ     
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ     
          в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ     
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ     

julia> sigmoid === Пғ
true

NNlib.sigmoid_fast вҖ” Function

sigmoid_fast(x)

This is a faster, and very slightly less accurate, version of sigmoid. For `x::Float32, perhaps 3 times faster, and maximum errors 2 eps instead of 1.

See also tanh_fast.

julia> sigmoid(0.2f0)
0.54983395f0

julia> sigmoid_fast(0.2f0)
0.54983395f0

julia> hardПғ(0.2f0)
0.53333336f0

NNlib.softplus вҖ” Function

softplus(x) = log(exp(x) + 1)

See "Deep Sparse Rectifier Neural Networks", JMLR 2011.

julia> lineplot(softplus, -3, 3, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ            
        4 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ softplus(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в”Ӯ            
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв Ғв Җв”Ӯ            
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв ”в Ҡв Ғв Җв Җв Җв Җв Җв”Ӯ            
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвЈҖвЎ в Өв ’в үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈҖвЎв Өв ’в Ҡв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЎ в Өв Өв Өв Өв ”в ’в ’в ҡв үв үв Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ            
          в Җ-3в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ3в Җ            
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ            

julia> lineplot!(ans, relu)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ            
        4 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ softplus(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈ в”Ӯ relu(x)    
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈ вЎҙв һв Ӣв Ғв”Ӯ            
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈӨвЎҙв һв Ӣв Ғв Җв Җв Җв Җв”Ӯ            
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв ҖвЈҖвЎ вўӨвЎІв қв Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈҖвЎв Өв ’в ҠвЈүв Ҙв ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈ вЈӨвЈӨвЈӨвЈӨвЈ”вЈ’вЈ’вЈҡвЈүвЈүвЈҒвЈҖвЈҮв ҙв ’в үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ            
          в Җ-3в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ3в Җ            
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ            

julia> softplus(16f0)
16.0f0

NNlib.softshrink вҖ” Function

softshrink(x, О»=0.5) =
    (x вүҘ О» ? x - О» : (-О» вүҘ x ? x + О» : 0))

See "Softshrink Activation Function".

julia> lineplot(softshrink, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ              
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв”Ӯ softshrink(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЎӨв ”в ’в үв Ғв”Ӯ              
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв Өв ’в Ӣв Ғв Җв Җв Җв Җв Җв Җв”Ӯ              
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвЈӨвЎӨв Өв Өв Өв Өв Өв ӨвЎв Өв Өв Өв Өв ¶в ®в ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ              
           в”Ӯв Җв Җв Җв Җв Җв ҖвўҖвЈҖв Өв –в ’в үв Ғв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в”Ӯв ҖвЈҖв Өв ”в ’в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
        -2 в”Ӯв үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ              
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ              
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ              

julia> lineplot!(ans, tanhshrink)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ              
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв”Ӯ softshrink(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЎӨв ”в ’вЈүвЎЎв”Ӯ tanhshrink(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв ӨвЈ’вЈӢв Ҙв Өв ’в Ҡв үв Ғв Җв”Ӯ              
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв ӨвЈӨвЈӨвЈӨвЈӨвЎӨв Өв Өв Өв Өв Өв ӨвЎ·в ¶в ¶в ¶в ¶в ¶в ҫв ҝв Ҝв ӯв ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ              
           в”Ӯв ҖвўҖвЈҖвЎ в Өв –вў’вЈӢв ӯв —в ’в үв Ғв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в”Ӯв ҠвЈүв Өв ”в ’в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
        -2 в”Ӯв үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ              
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ              
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җ

julia> softshrink.((-10f0, 10f0))
(-9.5f0, 9.5f0)

NNlib.softsign вҖ” Function

softsign(x) = x / (1 + |x|)

See "Quadratic Polynomials Learn Better Image Features" (2009).

julia> lineplot(softsign, -5, 5, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ            
         1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈҖвЈҖвЈҖвЈҖв Өв Өв Өв Өв Өв”Ӯ softsign(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЈҖвЎӨв –в ’в Ӣв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвЎ”в Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎҜв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ”в ҒвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈҖвЈҖв Өв Өв ’в Ӣв Ғв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
        -1 в”Ӯв ’в ’в ’в ’в ’в Ҡв үв үв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ            
           в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ            
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ            

julia> lineplot!(ans, tanh)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ            
         1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвўҖвЎӨв –в Ҡв үв үв үвЈүвЈүвЈүвЈүвЈүв ӯв ӯв ӯв ӯв ӯв”Ӯ softsign(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв ҖвЎ”вЈғвЎӨв –в ’в Ӣв үв үв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ tanh(x)    
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈвЎһв Ӣв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЎҜв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎҙв ғвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЈҖвЈҖв Өв Өв ’вўӢв •в Ғв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
        -1 в”ӮвЈ’вЈ’вЈ’вЈ’вЈ’вЈҠвЈүвЈүвЈүвЈүвЈҒвЈҖвЈҖвЎ в Өв ’в Ғв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ            
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ            
           в Җ-5в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ5в Җ            
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ            

julia> softsign(1f0)
0.5f0

julia> softsign(100f0)
0.990099f0

NNlib.swish вҖ” Function

swish(x) = x * Пғ(x)

Self-gated activation function. See "Swish: a Self-Gated Activation Function".

julia> lineplot(swish, -2, 2, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
         2 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв”Ӯ swish(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ӣв Ғв Җв”Ӯ         
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв Өв –в Ӣв Ғв Җв Җв Җв Җв Җв”Ӯ         
   f(x)    в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв ҖвўҖвЈҖвЎӨв ”в Ҡв үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
           в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЈӨвЈӨвЎӨвЎв ҙв ¶в Ҝв Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ         
           в”Ӯв үв ‘в ’в ’в ’в ’в ’в ’в ’в ’в ’в ’в үв үв үв үв Ғв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        -1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
           в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ2в Җ         
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

NNlib.tanhshrink вҖ” Function

tanhshrink(x) = x - tanh(x)

See "Tanhshrink Activation Function".

julia> lineplot(tanhshrink, -3, 3, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ              
         3 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ tanhshrink(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎ в Өв –в Ҡв”Ӯ              
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвўҖвЈҖвЎ в Өв ’в Ҡв үв Ғв Җв Җв Җв Җв”Ӯ              
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвўӨвЈӨвЎӨв Өв Өв Өв Өв Өв ӨвЎ·в ¶в ¶в ¶в ¶в ¶в ®в ӯв Ҙв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ              
           в”Ӯв Җв Җв Җв Җв ҖвЈҖвЎ в ҙв ’в Ҡв үв Ғв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в”ӮвЎ в ҙв ’в Ҡв үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
        -3 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ              
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ              
           в Җ-3в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ3в Җ              
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ              

julia> tanhshrink.((-10f0, 10f0))
(-9.0f0, 9.0f0)

NNlib.tanh_fast вҖ” Function

tanh_fast(x)

This is a faster but slighly less accurate version of tanh.

Where Julia's tanh function has an error under 2 eps, this may be wrong by 5 eps, a reduction by less than one decimal digit.

For x::Float32 this is usually about 10 times faster, with a smaller speedup for x::Float64. For any other number types, it just calls tanh.

See also sigmoid_fast.

julia> tanh(0.5f0)
0.46211717f0

julia> tanh_fast(0.5f0)
0.46211714f0

julia> hard_tanh(0.5f0)
0.5f0

NNlib.trelu вҖ” Function

trelu(x, theta=1) = x > theta ? x : 0

Threshold gated rectified linear activation function. See "Zero-bias autoencoders and the benefits of co-adapting features"

julia> lineplot(trelu, -2, 4, height=7)
          в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ         
        4 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖвЎӨв –в Ӣв”Ӯ trelu(x)
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в –в Ӣв Ғв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвўҖвЎ в ”в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
   f(x)   в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ҙв Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв ҖвЈ в Өв ’в үв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв ҖвЎҸв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
        0 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ         
          в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ         
          в Җ-2в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ4в Җ         
          в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ         

One More

Julia's Base.Math also provides tanh, which can be used as an activation function.

Note that many Flux layers will automatically replace this with NNlib.tanh_fast when called, as Base's tanh is slow enough to sometimes be a bottleneck.

julia> using UnicodePlots

julia> lineplot(tanh, -3, 3, height=7)
           в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ        
         1 в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв ҖвЈҖв Өв ”в ’в ’в үв үв үв үв үв үв үв үв үв”Ӯ tanh(x)
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮв Җв ҖвЎ в –в Ӣв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎҮвЎ°в Ҡв Ғв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
   f(x)    в”Ӯв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв ӨвЎӨвЎҜв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв Өв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЎ в Һв ҒвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в”Ӯв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв ҖвЈҖв ҙв Ҡв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
        -1 в”ӮвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЈҖвЎӨв Өв ”в ’в үв Ғв Җв Җв Җв Җв ҖвЎҮв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв”Ӯ        
           в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ        
           в Җ-3в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ3в Җ        
           в Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җxв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җв Җ