Activation Functions from NNlib.jl

These non-linearities used between layers of your model are exported by the NNlib package.

Note that, unless otherwise stated, activation functions operate on scalars. To apply them to an array you can call ¤Ѓ.(xs), relu.(xs) and so on. Alternatively, they can be passed to a layer like Dense(784 => 1024, relu) which will handle this broadcasting.

Functions like softmax are sometimes described as activation functions, but not by Flux. They must see all the outputs, and hence cannot be broadcasted. See the next page for details.

Alphabetical Listing

NNlib.celu Рђћ Function

celu(x, ╬▒=1) = x РЅЦ 0 ? x : ╬▒ * (exp(x/╬▒) - 1)

Activation function from "Continuously Differentiable Exponential Linear Units".

julia> lineplot(celu, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРањРаЅРћѓ celu(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЅРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРАцРаќРаіРаЂРађРађРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРБђРацРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРАцРАДРаХРаГРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРбђРБђРБђРацРаћРањРаІРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -1 РћѓРацРацРацРацРаћРањРањРањРаіРаЅРаЅРаЂРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

julia> celu(-10f0)
-0.9999546f0

NNlib.elu Рђћ Function

elu(x, ╬▒=1) = x > 0 ? x : ╬▒ * (exp(x) - 1)

Exponential Linear Unit activation function. See "Fast and Accurate Deep Network Learning by Exponential Linear Units". You can also specify the coefficient explicitly, e.g. elu(x, 1).

julia> lineplot(elu, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ       
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРањРаЅРћѓ elu(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЅРађРађРађРађРћѓ       
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРАцРаќРаіРаЂРађРађРађРађРађРађРађРађРћѓ       
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРБђРацРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
           РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРАцРАДРаХРаГРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ       
           РћѓРађРађРађРађРађРађРађРађРађРађРађРбђРБђРБђРацРаћРањРаІРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
        -1 РћѓРацРацРацРацРаћРањРањРањРаіРаЅРаЅРаЂРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў       
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ       
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ       

julia> elu(-10f0)
-0.9999546f0

julia> elu(-10f0, 2)
-1.9999092f0

NNlib.gelu Рђћ Function

gelu(x) = 0.5x * (1 + tanh(Рѕџ(2/¤ђ) * (x + 0.044715x^3)))

Activation function from "Gaussian Error Linear Units".

julia> lineplot(gelu, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРћѓ gelu(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаіРаЂРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРБђРацРањРаЅРађРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРБђРАаРацРањРаЅРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРБцРБцРБцРБцРБцРБцРБцРБцРАцРацРацРацРацРацРацРацРБцРБцРБцРАцРАДРаХРаХРаГРаЦРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ        
           РћѓРађРађРађРађРађРађРађРађРаѕРаЅРаЅРаЅРаЅРаЅРаЅРаЅРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

julia> lineplot(gelu, -5, 0, height=7);

julia> lineplot!(ans, swish)
             РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
           0 РћѓРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРањРањРацРБёРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРћѓ gelu(x) 
             РћѓРаЉРањРабРацРБёРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаЅРаЊРбёРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРћѓ swish(x)
             РћѓРађРађРађРађРађРаѕРаЅРањРацРБђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаЉРбєРАђРађРађРађРађРађРађРађРађРађРађРБИРаЂРћѓ         
   f(x)      РћѓРађРађРађРађРађРађРађРађРађРађРаЅРањРбёРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРаЉРбёРађРађРађРађРађРађРађРађРбаРАЄРађРћѓ         
             РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРаЅРаЊРбёРАђРађРађРађРађРађРађРађРађРађРађРађРађРаЊРБёРађРађРађРађРађРбаРАъРађРађРћѓ         
             РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаЅРадРАђРађРађРађРађРађРађРађРађРађРађРађРађРаЊРбёРБђРБђРАцРбБРаЃРађРађРћѓ         
        -0.2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаѕРаЊРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбаРаЄРађРађРађРћѓ         
             РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
             Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ0Рађ         
             РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

NNlib.hardsigmoid Рђћ Function

hard¤Ѓ(x) = max(0, min(1, (x + 3) / 6))

Piecewise linear approximation of sigmoid.

julia> lineplot(hardsigmoid, -5, 5, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
        1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРАаРаќРаІРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ hard¤Ѓ(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРБђРАцРањРаЂРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРАаРаћРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРАЌРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаіРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРбђРАцРаќРаІРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРацРаіРаЂРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
          Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ         
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

julia> lineplot(sigmoid, -5, 5, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ     
        1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРАаРацРаќРањРањРаІРаЅРаЅРаЅРаЅРаЅРаЅРћѓ ¤Ѓ(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРбђРАаРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРБђРаћРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАаРАЈРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАћРаІРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРаіРаЂРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРацРацРацРањРаіРаЅРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў     
          Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ     
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ     

NNlib.sigmoid_fast Рђћ Function

sigmoid_fast(x)

This is a faster, and very slightly less accurate, version of sigmoid. For `x::Float32, perhaps 3 times faster, and maximum errors 2 eps instead of 1.

See also tanh_fast.

julia> sigmoid(0.2f0)
0.54983395f0

julia> sigmoid_fast(0.2f0)
0.54983395f0

julia> hard¤Ѓ(0.2f0)
0.53333336f0

NNlib.hardtanh Рђћ Function

hardtanh(x) = max(-1, min(1, x))

Segment-wise linear approximation of tanh, much cheaper to compute. See "Large Scale Machine Learning".

See also tanh_fast.

julia> lineplot(hardtanh, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ            
         1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРаћРаІРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ hardtanh(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРБђРАцРаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРбђРАцРаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРАцРАиРаЦРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАаРаќРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАаРаќРаІРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
        -1 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРаћРаІРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў            
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ            
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађx

julia> lineplot(tanh, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРБђРАаРацРацРањРањРањРаіРаЅРаЅРаЅРћѓ tanh(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРбђРАаРаћРаіРаЅРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРбђРАцРањРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРАцРАиРаЦРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАцРаќРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРБђРАаРаћРаіРаЂРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -1 РћѓРБђРБђРБђРАаРацРацРацРаќРањРаіРаЅРаЂРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

NNlib.tanh_fast Рђћ Function

tanh_fast(x)

This is a faster but slighly less accurate version of tanh.

Where Julia's tanh function has an error under 2 eps, this may be wrong by 5 eps, a reduction by less than one decimal digit.

For x::Float32 this is usually about 10 times faster, with a smaller speedup for x::Float64. For any other number types, it just calls tanh.

See also sigmoid_fast.

julia> tanh(0.5f0)
0.46211717f0

julia> tanh_fast(0.5f0)
0.46211714f0

julia> hard_tanh(0.5f0)
0.5f0

NNlib.leakyrelu Рђћ Function

leakyrelu(x, a=0.01) = max(a*x, x)

Leaky Rectified Linear Unit activation function. You can also specify the coefficient explicitly, e.g. leakyrelu(x, 0.01).

julia> lineplot(x -> leakyrelu(x, 0.5), -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ       
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРањРаЅРћѓ #42(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЅРађРађРађРађРћѓ       
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРАцРаќРаіРаЂРађРађРађРађРађРађРађРађРћѓ       
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРБђРацРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
           РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРБцРБцРАцРАДРаХРаГРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ       
           РћѓРађРађРађРађРађРађРађРађРбђРБђРБђРацРацРањРањРаІРаЅРаЂРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
        -1 РћѓРБђРБђРацРацРањРањРаіРаЅРаЂРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ       
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў       
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ       
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ       

julia> leakyrelu(-10f0, 0.2)
-2.0f0

julia> leakyrelu(-10f0, 0.02)
-0.5f0

NNlib.lisht Рђћ Function

lisht(x) = x * tanh(x)

Activation function from "LiSHT: Non-Parametric Linearly Scaled Hyperbolic Tangent ..."

julia> lineplot(lisht, -2, 2, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
        2 РћѓРабРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРћѓ lisht(x)
          РћѓРађРаѕРаЉРбдРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАцРаіРаЂРађРћѓ         
          РћѓРађРађРађРађРаѕРаБРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаЂРађРађРађРађРћѓ         
   f(x)   РћѓРађРађРађРађРађРађРађРаЉРбєРАђРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРАаРаіРаЂРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРаЅРабРАёРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРаћРаІРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРаѕРаЊРбёРАђРађРађРађРађРађРАЄРађРађРађРађРбђРАаРаќРаЂРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        0 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаѕРаЊРадРБёРБђРБђРБЄРБђРБђРацРањРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
          Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ         
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

julia> lineplot!(ans, logcosh)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ           
        2 РћѓРабРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРћѓ lisht(x)  
          РћѓРађРаѕРаЉРбдРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАцРаіРаЂРађРћѓ logcosh(x)
          РћѓРабРБёРађРађРаѕРаБРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаЂРађРађРБђРаћРћѓ           
   f(x)   РћѓРађРаѕРаЉРабРБђРађРађРаЉРбєРАђРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРАаРаіРаЂРађРБђРаћРаіРаЂРађРћѓ           
          РћѓРађРађРађРађРађРаЅРабРбёРАђРаЅРабРАёРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРаћРаІРађРАаРаћРаІРаЂРађРађРађРађРћѓ           
          РћѓРађРађРађРађРађРађРађРађРаЅРаЊРадРБїРАЊРбёРАђРађРађРађРађРађРАЄРађРађРађРађРбђРАаРаќРБЂРацРањРаЅРађРађРађРађРађРађРађРађРћѓ           
        0 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРаЅРаЊРафРаиРБдРБёРБђРБђРБЄРБђРБђРБцРаХРаЋРањРаЅРађРађРађРађРађРађРађРађРађРађРађРађРћѓ           
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў           
          Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ           
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ           

NNlib.logcosh Рђћ Function

logcosh(x)

Return log(cosh(x)) which is computed in a numerically stable way.

julia> lineplot(logcosh, -5, 5, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ           
        5 РћѓРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ logcosh(x)
          РћѓРаЅРабРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаІРћѓ           
          РћѓРађРађРађРаЉРабРБёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаіРаЂРађРађРћѓ           
   f(x)   РћѓРађРађРађРађРађРађРаЉРадРБђРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРБђРАцРаіРаЂРађРађРађРађРађРћѓ           
          РћѓРађРађРађРађРађРађРађРађРађРаЉРадРАђРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРАцРаіРаЂРађРађРађРађРађРађРађРађРћѓ           
          РћѓРађРађРађРађРађРађРађРађРађРађРађРаѕРаЊРадРАђРађРађРађРађРађРАЄРађРађРађРађРбђРАцРањРаЂРађРађРађРађРађРађРађРађРађРађРађРћѓ           
        0 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаѕРаЉРабРбёРБђРБђРБЄРБђРАаРаћРаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ           
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў           
          Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ           
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ           

NNlib.logsigmoid Рђћ Function

log¤Ѓ(x)

Return log(¤Ѓ(x)) which is computed in a numerically stable way.

julia> lineplot(logsigmoid, -5, 5, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         0 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАДРацРаћРањРањРањРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ log¤Ѓ(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРаќРаіРаЅРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРБђРацРањРаЅРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРађРађРађРађРађРађРбђРАцРаќРаІРаЂРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРБђРаћРаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРАцРаќРаІРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -6 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

NNlib.mish Рђћ Function

mish(x) = x * tanh(softplus(x))

Activation function from "Mish: A Self Regularized Non-Monotonic Neural Activation Function".

julia> lineplot(mish, -5, 5, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         5 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРаќРаІРћѓ mish(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРањРаЂРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРАаРаћРаІРаЂРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРбђРАаРаќРаІРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРбђРАцРаќРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБДРБћРБіРБЂРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРћѓ        
        -1 РћѓРађРађРађРађРађРађРађРађРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

NNlib.relu Рђћ Function

relu(x) = max(0, x)

Rectified Linear Unit activation function.

julia> lineplot(relu, -2, 2, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
        2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаІРћѓ relu(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРаіРаЂРађРађРћѓ        
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРбђРАцРаіРаЂРађРађРађРађРађРћѓ        
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРАцРаќРаЂРађРађРађРађРађРађРађРађРћѓ        
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРАаРаќРаЂРађРађРађРађРађРађРађРађРађРађРађРћѓ        
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРАаРаќРаІРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБЄРаћРаІРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
          Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ        
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

NNlib.relu6 Рђћ Function

relu6(x) = min(max(0, x), 6)

Rectified Linear Unit activation function capped at 6. See "Convolutional Deep Belief Networks" from CIFAR-10.

julia> lineplot(relu6, -10, 10, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
        6 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРБђРајРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ relu6(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРАћРаЂРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРАцРаЃРађРађРађРађРађРађРађРађРађРађРађРћѓ         
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРАаРајРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРбђРаќРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРАћРаЃРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРАДРаІРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
          Рађ-10РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ10Рађ         
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

NNlib.rrelu Рђћ Function

rrelu(x, lo=1/8, hi=1/3) = max(a*x, x)
# where `a` is randomly sampled from uniform distribution `U(lo, hi)`

Randomized Leaky Rectified Linear Unit activation function. See "Empirical Evaluation of Rectified Activations" You can also specify the bound explicitly, e.g. rrelu(x, 0.0, 1.0).

julia> lineplot(rrelu, -20, 10, height=7)
            РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
         10 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРађРађРађРађРађРађРађРађРађРБђРАцРаќРаІРћѓ rrelu(x)
            РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРађРађРађРађРађРбђРАаРаќРаІРаЂРађРађРађРћѓ         
            РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРађРбђРАаРаћРаіРаЂРађРађРађРађРађРађРађРћѓ         
   f(x)     РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРАцРацРБцРБцРбцРБцРБцРацРацРацРб╝Ра«РаЦРацРацРацРацРацРацРацРацРацРацРацРћѓ         
            РћѓРБ░РбђРБєРАёРБёРАёРАаРА░РадРаиРАюРббРаиРа│РабРаіРаЅРаЅРађРађРаЂРађРађРађРађРађРбИРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
            РћѓРаЃРаЅРаЎРаўРаЃРаѕРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        -10 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбИРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
            РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
            Рађ-20РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ10Рађ         
            РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

julia> extrema(rrelu.(fill(-10f0, 1000)))
(-3.3316886f0, -1.2548422f0)

NNlib.selu Рђћ Function

selu(x) = ╬╗ * (x РЅЦ 0 ? x : ╬▒ * (exp(x) - 1))

╬╗ РЅѕ 1.05070...
╬▒ РЅѕ 1.67326...

Scaled exponential linear units. See "Self-Normalizing Neural Networks".

julia> lineplot(selu, -3, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         3 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ selu(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРБђРАцРацРањРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРбђРБђРацРаќРаіРаЅРађРађРађРађРћѓ        
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРБђРАаРацРањРаІРаЂРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРБЅРаГРаЏРАЈРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРБђРБђРАцРацРањРаіРаЅРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -2 РћѓРацРацРаќРањРањРањРањРањРањРањРаЅРаЅРаЅРаЂРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-3РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

julia> selu(-10f0)
-1.7580194f0

NNlib.sigmoid Рђћ Function

¤Ѓ(x) = 1 / (1 + exp(-x))

Classic sigmoid activation function. Unicode ¤Ѓ can be entered as \sigma then tab, in many editors. The ascii name sigmoid is also exported.

See also sigmoid_fast.

julia> using UnicodePlots

julia> lineplot(sigmoid, -5, 5, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ     
        1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРБђРАаРацРаќРањРањРаІРаЅРаЅРаЅРаЅРаЅРаЅРћѓ ¤Ѓ(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРбђРАаРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРБђРаћРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАаРАЈРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАћРаІРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРаіРаЂРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРацРацРацРањРаіРаЅРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ     
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў     
          Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ     
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ     

julia> sigmoid === ¤Ѓ
true

NNlib.softplus Рђћ Function

softplus(x) = log(exp(x) + 1)

See "Deep Sparse Rectifier Neural Networks", JMLR 2011.

julia> lineplot(softplus, -3, 3, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ            
        4 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ softplus(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРћѓ            
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЂРађРћѓ            
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРБђРАцРаћРаіРаЂРађРађРађРађРађРћѓ            
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРБђРАаРацРањРаЅРаЂРађРађРађРађРађРађРађРађРађРћѓ            
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРБђРАДРацРањРаіРаЅРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРАаРацРацРацРацРаћРањРањРаџРаЅРаЅРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў            
          Рађ-3РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ3Рађ            
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ            

julia> lineplot!(ans, relu)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ            
        4 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ softplus(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРБаРћѓ relu(x)    
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРБаРА┤РаъРаІРаЂРћѓ            
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРБђРБцРА┤РаъРаІРаЂРађРађРађРађРћѓ            
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРБђРАаРбцРА▓РаЮРаІРаЂРађРађРађРађРађРађРађРађРћѓ            
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРБђРАДРацРањРаіРБЅРаЦРаџРаЂРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРБаРБцРБцРБцРБцРБћРБњРБњРБџРБЅРБЅРБЂРБђРБЄРа┤РањРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў            
          Рађ-3РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ3Рађ            
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ            

julia> softplus(16f0)
16.0f0

NNlib.softshrink Рђћ Function

softshrink(x, ╬╗=0.5) =
    (x РЅЦ ╬╗ ? x - ╬╗ : (-╬╗ РЅЦ x ? x + ╬╗ : 0))

See "Softshrink Activation Function".

julia> lineplot(softshrink, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ              
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРћѓ softshrink(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРбђРБђРАцРаћРањРаЅРаЂРћѓ              
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРБђРАцРацРањРаІРаЂРађРађРађРађРађРађРћѓ              
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРБцРАцРацРацРацРацРацРацРАДРацРацРацРацРаХРа«РаГРацРацРацРацРацРацРацРацРацРацРацРацРћѓ              
           РћѓРађРађРађРађРађРађРбђРБђРацРаќРањРаЅРаЂРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћѓРађРБђРацРаћРањРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
        -2 РћѓРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў              
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ              
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ              

julia> lineplot!(ans, tanhshrink)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ              
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРћѓ softshrink(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРбђРБђРАцРаћРањРБЅРААРћѓ tanhshrink(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРБђРАцРацРБњРБІРаЦРацРањРаіРаЅРаЂРађРћѓ              
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРБцРБцРБцРБцРАцРацРацРацРацРацРацРАиРаХРаХРаХРаХРаХРаЙРа┐Ра»РаГРаГРацРацРацРацРацРацРацРацРацРћѓ              
           РћѓРађРбђРБђРАаРацРаќРбњРБІРаГРаЌРањРаЅРаЂРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћѓРаіРБЅРацРаћРањРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
        -2 РћѓРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў              
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ              
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађ

julia> softshrink.((-10f0, 10f0))
(-9.5f0, 9.5f0)

NNlib.softsign Рђћ Function

softsign(x) = x / (1 + |x|)

See "Quadratic Polynomials Learn Better Image Features" (2009).

julia> lineplot(softsign, -5, 5, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ            
         1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРБђРБђРБђРБђРБђРацРацРацРацРацРћѓ softsign(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРБђРАцРаќРањРаІРаЅРаЅРаЂРађРађРађРађРађРађРађРађРађРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРАћРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРА»РацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРаћРаЂРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРБђРБђРБђРацРацРањРаІРаЂРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
        -1 РћѓРањРањРањРањРањРаіРаЅРаЅРаЅРаЅРаЂРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў            
           Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ            
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ            

julia> lineplot!(ans, tanh)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ            
         1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРбђРАцРаќРаіРаЅРаЅРаЅРБЅРБЅРБЅРБЅРБЅРаГРаГРаГРаГРаГРћѓ softsign(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРАћРБЃРАцРаќРањРаІРаЅРаЅРаЂРађРађРађРађРађРађРађРађРађРћѓ tanh(x)    
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБДРАъРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРА»РацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРА┤РаЃРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћѓРађРађРађРађРађРађРађРађРађРађРБђРБђРБђРацРацРањРбІРаЋРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
        -1 РћѓРБњРБњРБњРБњРБњРБіРБЅРБЅРБЅРБЅРБЂРБђРБђРАаРацРањРаЂРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ            
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў            
           Рађ-5РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ            
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ            

julia> softsign(1f0)
0.5f0

julia> softsign(100f0)
0.990099f0

NNlib.swish Рђћ Function

swish(x) = x * ¤Ѓ(x)

Self-gated activation function. See "Swish: a Self-Gated Activation Function".

julia> lineplot(swish, -2, 2, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
         2 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРћѓ swish(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРаќРаІРаЂРађРћѓ         
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРБђРацРаќРаІРаЂРађРађРађРађРађРћѓ         
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРбђРБђРАцРаћРаіРаЅРађРађРађРађРађРађРађРађРађРађРћѓ         
           РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРбцРБцРБцРАцРАДРа┤РаХРа»РаЦРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ         
           РћѓРаЅРаЉРањРањРањРањРањРањРањРањРањРањРаЅРаЅРаЅРаЅРаЂРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        -1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ2Рађ         
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

NNlib.hardswish Рђћ Function

hardswish(x) = x * hard¤Ѓ(x)

Hard-Swish activation function. See "Searching for MobileNetV3".

julia> lineplot(hardswish, -2, 5, height = 7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ             
         5 РћѓРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРаћРањРаЅРћѓ hardswish(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАаРаћРањРаЅРаЂРађРађРађРађРћѓ             
           РћѓРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРацРаќРаЅРаЂРађРађРађРађРађРађРађРађРађРћѓ             
   f(x)    РћѓРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ             
           РћѓРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРбђРБђРацРаќРаІРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ             
           РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБЄРБцРБцРБќРБџРБЅРБЂРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРћѓ             
        -1 РћѓРаЅРањРањРањРањРаЅРаЅРаЅРаЅРаЂРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ             
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў             
           Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ5Рађ             
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ             

julia> lineplot(hardswish, -4, 0, height = 7);

julia> lineplot!(ans, swish)
             РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ             
           0 РћѓРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРбБРАђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАюРћѓ hardswish(x)
             РћѓРањРањРабРацРбёРБђРАђРађРађРађРађРа▒РАёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРајРађРћѓ swish(x)    
             РћѓРађРађРађРађРађРађРаѕРаЅРаЉРањРадРбёРБўРбёРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРА┤РаЃРађРађРћѓ             
   f(x)      РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРаЅРаЉРАќРадРбёРБђРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРбћРаЈРаЂРађРађРађРћѓ             
             РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаѕРаБРБёРађРаЅРаЉРањРадРацРбёРБђРБђРБђРБђРАаРацРаќРБіРаЋРаЂРађРађРађРађРађРћѓ             
             РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаЊРацРАђРађРађРађРађРађРађРађРађРбђРАцРаќРаЂРађРађРађРађРађРађРађРћѓ             
        -0.4 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРаѕРаЅРањРабРацРацРаћРањРаІРаЂРађРађРађРађРађРађРађРађРађРађРћѓ             
             РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў             
             Рађ-4РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ0Рађ             
             РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ             

julia> hardswish.(-5:5)'
1├Ќ11 adjoint(::Vector{Float64}) with eltype Float64:
 -0.0  -0.0  -0.0  -0.333333  -0.333333  0.0  0.666667  1.66667  3.0  4.0  5.0

NNlib.tanhshrink Рђћ Function

tanhshrink(x) = x - tanh(x)

See "Tanhshrink Activation Function".

julia> lineplot(tanhshrink, -3, 3, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ              
         3 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ tanhshrink(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАаРацРаќРаіРћѓ              
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРбђРБђРАаРацРањРаіРаЅРаЂРађРађРађРађРћѓ              
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРбцРБцРАцРацРацРацРацРацРацРАиРаХРаХРаХРаХРаХРа«РаГРаЦРацРацРацРацРацРацРацРацРацРацРацРћѓ              
           РћѓРађРађРађРађРађРБђРАаРа┤РањРаіРаЅРаЂРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћѓРАаРа┤РањРаіРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
        -3 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ              
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў              
           Рађ-3РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ3Рађ              
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ              

julia> tanhshrink.((-10f0, 10f0))
(-9.0f0, 9.0f0)

NNlib.trelu Рђћ Function

trelu(x, theta=1) = x > theta ? x : 0

Threshold gated rectified linear activation function. See "Zero-bias autoencoders and the benefits of co-adapting features"

julia> lineplot(trelu, -2, 4, height=7)
          РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ         
        4 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРАцРаќРаІРћѓ trelu(x)
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРаќРаІРаЂРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРбђРАаРаћРаіРаЂРађРађРађРађРађРађРађРћѓ         
   f(x)   РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРБђРа┤РаіРаЂРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРађРБаРацРањРаЅРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРађРАЈРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
        0 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБђРБЄРБђРБђРБђРБђРБђРБђРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ         
          РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў         
          Рађ-2РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ4Рађ         
          РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ         

One More

Julia's Base.Math also provides tanh, which can be used as an activation function.

Note that many Flux layers will automatically replace this with NNlib.tanh_fast when called, as Base's tanh is slow enough to sometimes be a bottleneck.

julia> using UnicodePlots

julia> lineplot(tanh, -3, 3, height=7)
           РћїРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ        
         1 РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРађРађРађРБђРацРаћРањРањРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРаЅРћѓ tanh(x)
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРађРађРАаРаќРаІРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАЄРА░РаіРаЂРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
   f(x)    РћѓРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРАцРА»РацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРацРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРАаРајРаЂРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћѓРађРађРађРађРађРађРађРађРађРађРађРађРађРађРБђРа┤РаіРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
        -1 РћѓРБђРБђРБђРБђРБђРБђРБђРБђРБђРАцРацРаћРањРаЅРаЂРађРађРађРађРађРАЄРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРћѓ        
           РћћРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў        
           Рађ-3РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ3Рађ        
           РађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађxРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађРађ        

Base.tanh Рђћ Function

tanh(x)

Compute hyperbolic tangent of x.